types: data structures and behavior for module
This provides data structures for connections, and buckets that are
accessed through connections.  The connections signal whether they're
ready or not, allowing the application to determine whether to submit
data to a bucket or retain it locally.

Buckets include information about the measurement structures, and
goroutines to handle communication with the server in the background.
The application is notified of the success or failure of every server
communication.  all failed writes will be offered to the user to
determine whether to resubmit, drop, or cache locally, and on shutdown
all unsent and in-flight data is reclaimed and presented to the
application for disposition.  Unlike the upstream WriteAPI this
implementation guarantees that no data will be dropped without giving
the application an opportunity to record it for future resubmission.

All operations traffic in newline-terminated Line Protocol-encoded
batches, simplifying local storage of data when the server is
unavailable.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>